Merges Fixes from Main & Adds Comment Q's
Sailor pushed a version of the code in the "main" branch. Migrated what I could (was not able to do it all) and started to review the code while adding in comments around sections that have me genuinely confused on their operation. Will convert to issues in the near future for information preservation.

Co-Authored-By: Robert-Abel Faustino Porras <24641567+MajorPing5@users.noreply.github.com>